JUCE: update to 8.0.13

+++ Source/Track/AnlTrackThumbnail.cpp
@@ -105,15 +105,15 @@ Track::Thumbnail::Thumbnail(Director& director, Zoom::Accessor& timeZoomAccessor
         {
             case SignalType::showProperties:
             {
-                auto const x = static_cast<int>(value.getProperty("x", 0.0));
-                auto const y = static_cast<int>(value.getProperty("y", 0.0));
+                auto const x = static_cast<float>(value.getProperty("x", 0.0));
+                auto const y = static_cast<float>(value.getProperty("y", 0.0));
                 mPropertyWindowContainer.showAt({x, y});
             }
             break;
             case SignalType::showTable:
             {
-                auto const x = static_cast<int>(value.getProperty("x", 0.0));
-                auto const y = static_cast<int>(value.getProperty("y", 0.0));
+                auto const x = static_cast<float>(value.getProperty("x", 0.0));
+                auto const y = static_cast<float>(value.getProperty("y", 0.0));
                 mResultsWindowContainer.showAt({x, y});
             }
             break;
